requirements.txt

# need cuda and cudatoolkit
# CUDAToolkit==12.1
# CUDA==12.1

openai==0.28.0
opencv-python==4.8.1.78
# opencv-contrib-python
Pillow==9.4.0
requests==2.29.0
requests-oauthlib==1.3.1
numpy==1.25.0
ptwt==0.1.7
pyautogui==0.9.54


